# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karag API",
    "description": "Modular RAG & Agentic Chatbot API",
    "contact": {
      "name": "Karag Team",
      "url": "https://github.com/karag-ai/karag"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/admin/baas/workspaces/{workspace_id}/api-keys": {
      "post": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for a workspace.\n\nSECURITY: This is the ONLY time the full key is returned.\nStore it securely - it cannot be retrieved later.\n\nRequired Permission: admin",
        "operationId": "create_api_key_api_v1_admin_baas_workspaces__workspace_id__api_keys_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "permissions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "read",
                "write"
              ],
              "title": "Permissions"
            }
          },
          {
            "name": "expires_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 365,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expires Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_APIKeyCreateResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "List Workspace Api Keys",
        "description": "List API keys for a workspace.\n\nNote: Returns metadata only, NEVER the full key.\n\nRequired Permission: admin",
        "operationId": "list_workspace_api_keys_api_v1_admin_baas_workspaces__workspace_id__api_keys_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_APIKey__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/baas/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Revoke Api Key",
        "description": "Revoke an API key.\n\nRequired Permission: admin",
        "operationId": "revoke_api_key_api_v1_admin_baas_api_keys__key_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/baas/system/config": {
      "get": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Get System Config",
        "description": "Get system-level configuration.\n\nRequired Permission: admin",
        "operationId": "get_system_config_api_v1_admin_baas_system_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_SystemConfig_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Update System Config",
        "description": "Update system configuration.\n\nSome changes may require a system restart.\n\nRequired Permission: admin",
        "operationId": "update_system_config_api_v1_admin_baas_system_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Updates"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_SystemConfig_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/baas/workspaces/{workspace_id}/config": {
      "get": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Get Workspace Config",
        "description": "Get workspace configuration.\n\nRequired Permission: admin",
        "operationId": "get_workspace_config_api_v1_admin_baas_workspaces__workspace_id__config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_WorkspaceConfig_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Update Workspace Config",
        "description": "Update workspace configuration.\n\nChanges are clamped to system limits.\n\nRequired Permission: admin",
        "operationId": "update_workspace_config_api_v1_admin_baas_workspaces__workspace_id__config_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Updates"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_WorkspaceConfig_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/baas/usage/logs": {
      "get": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Query Usage Logs",
        "description": "Query usage logs with filters.\n\nRequired Permission: admin",
        "operationId": "query_usage_logs_api_v1_admin_baas_usage_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "endpoint",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Endpoint"
            }
          },
          {
            "name": "status_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_UsageLog__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/baas/usage/workspaces/{workspace_id}/stats": {
      "get": {
        "tags": [
          "BaaS Admin"
        ],
        "summary": "Get Workspace Usage Stats",
        "description": "Get usage statistics for a workspace.\n\nRequired Permission: admin",
        "operationId": "get_workspace_usage_stats_api_v1_admin_baas_usage_workspaces__workspace_id__stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 1,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_WorkspaceUsageStats_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIKeyCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "FULL KEY - SAVE THIS NOW"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "description": "Security warning",
            "default": "This is the ONLY time the API key will be displayed. Store it securely."
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "api_key",
          "key_prefix",
          "permissions",
          "expires_at",
          "created_at"
        ],
        "title": "APIKeyCreateResponse",
        "description": "Response when creating a new API key - ONLY time full key is visible."
      },
      "AppResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse",
        "description": "Standardized API response wrapper.\n\nAll API responses follow this structure for consistency:\n{\n    \"success\": bool,\n    \"code\": str,\n    \"message\": str,\n    \"data\": T | null\n}\n\nType Parameters:\n    T: The type of data contained in the response\n\nExamples:\n    >>> # Success response with data\n    >>> response = AppResponse.success_response(\n    ...     data={\"id\": 1, \"name\": \"Test\"},\n    ...     message=\"User created successfully\"\n    ... )\n\n    >>> # Error response\n    >>> response = AppResponse.business_failure(\n    ...     code=\"VALIDATION_ERROR\",\n    ...     message=\"Invalid input provided\"\n    ... )"
      },
      "AppResponse_APIKeyCreateResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIKeyCreateResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[APIKeyCreateResponse]"
      },
...
      "AppResponse_SystemConfig_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SystemConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[SystemConfig]"
      },
...
      "AppResponse_WorkspaceConfig_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[WorkspaceConfig]"
      },
...
      "AppResponse_WorkspaceUsageStats_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceUsageStats"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[WorkspaceUsageStats]"
      },
...
      "AppResponse_list_APIKey__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/APIKey"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[list[APIKey]]"
      },
...
      "AppResponse_list_UsageLog__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UsageLog"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[list[UsageLog]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceConfig": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "enabled_datasets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enabled Datasets"
          },
          "default_dataset_id": {
            "type": "string",
            "title": "Default Dataset Id",
            "default": "default"
          },
          "rag_config": {
            "$ref": "#/components/schemas/RAGConfig"
          },
          "default_temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Default Temperature",
            "default": 0.7
          },
          "default_max_tokens": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Default Max Tokens",
            "default": 1024
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "default": "gpt-4o"
          },
          "rate_limits": {
            "$ref": "#/components/schemas/RateLimitConfig"
          }
        },
        "type": "object",
        "required": [
          "workspace_id"
        ],
        "title": "WorkspaceConfig",
        "description": "Workspace-level configuration (restricted).\n\nWorkspaces can modify these within system-defined limits."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}